+++ dwarfeng-datamark-api/src/main/java/module-info.java
@@ -1,11 +1,11 @@
 module com.dwarfeng.datamark.api {
 
     requires com.dwarfeng.datamark.base;
-    requires transitive com.dwarfeng.datamark.core;
-    requires transitive com.dwarfeng.springtelqos.core;
+    requires com.dwarfeng.datamark.core;
+    requires com.dwarfeng.springtelqos.core;
     requires org.apache.commons.cli;
     requires org.apache.commons.lang3;
-    requires static org.jetbrains.annotations;
+    requires org.jetbrains.annotations;
 
     exports com.dwarfeng.datamark.api.integration.springtelqos;